+++ apps/native-component-list/src/screens/Calendar@Next/EventsNextScreen.tsx
@@ -1,6 +1,6 @@
 import { StackScreenProps } from '@react-navigation/stack';
 import * as Calendar from 'expo-calendar';
-import { AddEventWithFormOptions, ExpoCalendar, ExpoCalendarEvent } from 'expo-calendar/next';
+import type { AddEventWithFormOptions, ExpoCalendar, ExpoCalendarEvent } from 'expo-calendar/next';
 import React, { useState, useEffect } from 'react';
 import { Alert, Platform, ScrollView, StyleSheet, Text, View } from 'react-native';
 
@@ -228,14 +228,23 @@ const EventsScreen = ({ route }: Props) => {
         <Button onPress={() => addEvent(true)} title="Add New Recurring Event" />
         <Button
           onPress={async () => {
-            const { calendar } = route.params!;
-            const newEvent = prepareEvent(calendar.id, true);
-            const result = await calendar.addEventWithForm(newEvent as AddEventWithFormOptions);
-            setTimeout(() => {
-              Alert.alert('createEventInCalendarAsync result', JSON.stringify(result), undefined, {
-                cancelable: true,
-              });
-            }, 200);
+            try {
+              const { calendar } = route.params!;
+              const newEvent = prepareEvent(calendar.id, true);
+              const result = await calendar.addEventWithForm(newEvent as AddEventWithFormOptions);
+              setTimeout(() => {
+                Alert.alert(
+                  'createEventInCalendarAsync result',
+                  JSON.stringify(result),
+                  undefined,
+                  {
+                    cancelable: true,
+                  }
+                );
+              }, 200);
+            } catch (e: any) {
+              Alert.alert('Error creating event with OS dialog', e.message);
+            }
           }}
           title="Create Event using the OS dialog"
         />