+++ packages/expo-calendar/src/next/Calendar.ts
@@ -289,14 +289,26 @@ export const getCalendarPermissions = InternalExpoCalendar.getCalendarPermission
 /**
  * Asks the user to grant permissions for accessing user's reminders.
  * @return A promise that resolves to an object of type [`PermissionResponse`](#permissionresponse).
+ * @platform ios
  */
-export const requestRemindersPermissions = InternalExpoCalendar.requestRemindersPermissions;
+export async function requestRemindersPermissions(): Promise<PermissionResponse> {
+  if (!InternalExpoCalendar.requestRemindersPermissions) {
+    throw new UnavailabilityError('Calendar', 'requestRemindersPermissions');
+  }
+  return await InternalExpoCalendar.requestRemindersPermissions();
+}
 
 /**
  * Checks user's permissions for accessing user's reminders.
  * @return A promise that resolves to an object of type [`PermissionResponse`](#permissionresponse).
+ * @platform ios
  */
-export const getRemindersPermissions = InternalExpoCalendar.getRemindersPermissions;
+export async function getRemindersPermissions(): Promise<PermissionResponse> {
+  if (!InternalExpoCalendar.getRemindersPermissions) {
+    throw new UnavailabilityError('Calendar', 'getRemindersPermissions');
+  }
+  return InternalExpoCalendar.getRemindersPermissions();
+}
 
 /**
  * Gets an array of Source objects with details about the different sources stored on the device.
@@ -375,6 +387,7 @@ export const useCalendarPermissions = createPermissionHook<
  * ```ts
  * const [status, requestPermission] = Calendar.useRemindersPermissions();
  * ```
+ * @platform ios
  */
 export const useRemindersPermissions = createPermissionHook({
   getMethod: getRemindersPermissions,